+++ common/src/slice/mod.rs
@@ -7,7 +7,7 @@ use nalgebra::{Vector2, Vector3};
 
 mod config;
 pub mod format;
-pub use config::{ExposureConfig, ExposureRemap, SliceConfig, Supersample};
+pub use config::{ExposureConfig, ExposureRemap, Height, SliceConfig, Supersample};
 pub use format::SliceMode;
 
 use crate::{