+++ mslicer/src/project/storage.rs
@@ -23,11 +23,15 @@ use slicer::{
 /// changes, even in development so anyone using the dev/prerelease versions
 /// don't ruin their projects.
 ///
-/// ## v14 (v?.?.?)
+/// ## v15 (v0.10.0)
+/// Store first layers and transition layers as either a layer count or a united
+/// distance.
+///
+/// ## v14 (v0.10.0)
 /// Separate XY and Z supersample antialiasing parameters. You might not always
 /// want 3DAA or the (honestly slight) performance impact.
 ///
-/// ## v13 (v?.?.?)
+/// ## v13 (v0.10.0)
 /// Merge exposure config lift and retract distances. Since these values really
 /// should always be the same.
 ///
@@ -66,7 +70,7 @@ use slicer::{
 /// ## v2 (v0.5.0)
 /// A complete rewrite using a custom serilizer/deserilizer because of the
 /// bincode drama...
-const VERSION: u16 = 14;
+const VERSION: u16 = 15;
 
 struct ModelInfo {
     mesh: u32,