+++ Changelog.md
@@ -55,6 +55,8 @@
 - Prevent opening the same tool window multiple times
 - Allow configuring quick layout min spacing
 - Remote print initialize at startup option now runs an initial scan
+- Rename 'First Layers' to 'Bottom Layers'
+- Allow setting slice config layer counts in millimeters
 
 ## v0.9.2 &mdash; August 30th, 2026